# backend/src/main/kotlin/com/fairvalue/web/PricingController.kt
package com.fairvalue.web

import com.fairvalue.domain.InstrumentType
import com.fairvalue.domain.JobStatus
import com.fairvalue.dto.JobContextDto
import com.fairvalue.dto.JobDto
import com.fairvalue.dto.JobListResponse
import com.fairvalue.dto.PriceJobResponse
import com.fairvalue.dto.PricingTrigger
import com.fairvalue.security.AuthPrincipal
import com.fairvalue.service.JobService
import com.fasterxml.jackson.databind.JsonNode
import org.springframework.http.HttpStatus
import org.springframework.security.core.annotation.AuthenticationPrincipal
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController
import java.time.LocalDate

/**
 * openapi: POST /instruments/{id}/price(trigger), GET /jobs/{job_id}, GET /jobs/{job_id}/result.
 * 실행은 VALUATOR+(서비스에서 강제), 조회는 인증된 전원. 모두 org_id 격리.
 */
@RestController
class PricingController(private val jobService: JobService) {

    @PostMapping("/instruments/{id}/price")
    @ResponseStatus(HttpStatus.CREATED)
    fun price(
        @AuthenticationPrincipal caller: AuthPrincipal,
        @PathVariable id: Long,
        @RequestBody(required = false) trigger: PricingTrigger?,
    ): PriceJobResponse = jobService.price(caller, id, trigger ?: PricingTrigger())

    // Phase 5-5: 평가 이력 목록(org 격리, 최신순, 필터 instrument_type/status/from/to).
    @GetMapping("/jobs")
    fun listJobs(
        @AuthenticationPrincipal caller: AuthPrincipal,
        @RequestParam(name = "instrument_type", required = false) type: InstrumentType?,
        @RequestParam(required = false) status: JobStatus?,
        @RequestParam(required = false) from: LocalDate?,
        @RequestParam(required = false) to: LocalDate?,
    ): JobListResponse = JobListResponse(items = jobService.listJobs(caller, type, status, from, to))

    @GetMapping("/jobs/{jobId}")
    fun getJob(
        @AuthenticationPrincipal caller: AuthPrincipal,
        @PathVariable jobId: Long,
    ): JobDto = jobService.getJob(caller, jobId)

    @GetMapping("/jobs/{jobId}/result")
    fun getResult(
        @AuthenticationPrincipal caller: AuthPrincipal,
        @PathVariable jobId: Long,
    ): JsonNode = jobService.getResult(caller, jobId)

    // ★5-7: 평가시점 입력 스냅샷(계산근거 '입력 스냅샷'). org 격리 동일. 구 job 은 has_context=false.
    @GetMapping("/jobs/{jobId}/context")
    fun getContext(
        @AuthenticationPrincipal caller: AuthPrincipal,
        @PathVariable jobId: Long,
    ): JobContextDto = jobService.getContext(caller, jobId)
}