# backend/src/main/resources/openapi.yaml
...
openapi: 3.1.0
info:
  title: FairValue Engine API
  version: 0.1.0
  description: >-
    복합금융상품·파생상품 공정가치 평가 플랫폼의 백엔드 REST API 계약(v0.1).
    7개 상품(RCPS·CPS·CB·EB·BW·SO·CSO)을 동일 인터페이스로 다룬다.
    모든 응답은 호출자의 조직(organization_id)으로 격리된다.
servers:
  - url: /api/v1
    description: 기본 API 베이스 경로
security:
  - bearerAuth: []
...
paths:
...
  /instruments/{id}/price:
    post:
      tags: [Pricing]
      summary: 평가 실행(trigger)
      description: >-
        trigger 방식. 본문에 rawForm 전체를 싣지 않는다. 서버가 저장된 terms(rawForm)를
        Normalizer 로 resolve 해 ValuationContext 를 내부 조립하고 input_hash 로 캐시를 판정한다.
        동일 input_hash 가 있으면 cached=true 로 기존 결과를 재사용한다.
      operationId: triggerPricing
      x-roles: [VALUATOR, ORG_ADMIN]
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PricingTrigger' }
      responses:
        '201':
          description: Job 생성(또는 캐시 적중)
          content:
            application/json:
              schema:
                type: object
                required: [job_id, status, cached]
                properties:
                  job_id: { type: integer }
                  status: { $ref: '#/components/schemas/JobStatus' }
                  cached: { type: boolean }
        '409':
          description: 입력 미완료/검증 실패로 실행 불가
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /jobs/{job_id}:
    get:
      tags: [Pricing]
      summary: Job 상태 조회
      operationId: getJob
      x-roles: [all]
      parameters:
        - $ref: '#/components/parameters/PathJobId'
      responses:
        '200':
          description: Job 상태
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
  /jobs/{job_id}/result:
    get:
      tags: [Pricing]
      summary: 평가 결과(PricingResult)
      description: '응답은 shared/schemas/pricing-result.schema.json 의 PricingResult 와 1:1.'
      operationId: getJobResult
      x-roles: [all]
      parameters:
        - $ref: '#/components/parameters/PathJobId'
      responses:
        '200':
          description: PricingResult
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PricingResult' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 아직 계산 미완료(status != DONE)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /jobs/{job_id}/context:
    get:
      tags: [Pricing]
      summary: 평가시점 입력 스냅샷(contextJson) — 5-7 계산근거 '입력 스냅샷'(additive)
      description: '평가 실행 시 resolve 된 ValuationContext 스냅샷. 구 job 은 has_context=false·context=null. org 격리 동일.'
      operationId: getJobContext
      x-roles: [all]
      parameters:
        - $ref: '#/components/parameters/PathJobId'
      responses:
        '200':
          description: 입력 스냅샷
          content:
            application/json:
              schema:
                type: object
                required: [job_id, has_context]
                properties:
                  job_id: { type: integer }
                  has_context: { type: boolean }
                  context: { type: [object, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <JWT>. 클레임의 organization_id 로 조직 격리.'
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: integer }
    PathJobId:
      in: path
      name: job_id
      required: true
      schema: { type: integer }
  responses:
...
    NotFound:
      description: 리소스 없음(조직 격리 포함)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnprocessableEntity:
      description: 검증 실패(필드 오류 포함)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      description: 공통 에러 본문.
      required: [code, message]
      properties:
        code: { type: string, description: '예: E001, E422 등' }
        message: { type: string }
        fields:
          type: array
          description: 필드별 검증 오류.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }

    # 출처: shared/schemas/pricing-result.schema.json#/$defs/issue
    Issue:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        field: { type: [string, 'null'] }
        stage: { type: [string, 'null'] }

    # ---- enum (출처: shared/schemas/instrument-types.ts) ----
    InstrumentType:
      type: string
      enum: [RCPS, CPS, CB, EB, BW, SO, CSO]
    ModelName:
      type: string
      description: 'TRINOMIAL 없음(미사용). GS=Goldman-Sachs 전환확률 가중 할인(엔진확장-2, additive).'
      enum: [TF_LATTICE, LATTICE, GS, MONTE_CARLO, LSMC, BSM, DCF]
...
    JobStatus:
      type: string
      enum: [QUEUED, RUNNING, DONE, FAILED, PARTIAL]
...
    EngineOptions:
      type: object
      properties:
        lattice_steps: { type: integer }
        simulation_paths: { type: integer }
        variance_reduction: { type: string, enum: [none, antithetic, sobol] }
    PricingTrigger:
      type: object
      description: >-
        평가 실행 trigger. rawForm 전체를 싣지 않는다. 저장된 terms 를 서버가 resolve 한다.
        params_override 는 저장값을 일시 덮어쓰는 제한적 override 용(예: 시드/패스수/단일 파라미터).
      properties:
        model: { $ref: '#/components/schemas/ModelName' }
        seed: { type: integer }
        options: { $ref: '#/components/schemas/EngineOptions' }
        params_override:
          type: object
          additionalProperties: true
          description: 제한적 override(저장 rawForm 의 일부 수치만). 전체 rawForm 대체 아님.
    Job:
      type: object
      required: [job_id, instrument_id, status]
      properties:
        job_id: { type: integer }
        instrument_id: { type: integer }
        status: { $ref: '#/components/schemas/JobStatus' }
        progress: { type: [number, 'null'], description: '0~1' }
        cached: { type: boolean }

    # ---- Curve/Asset ----
...
    PricingResult:
      type: object
      additionalProperties: false
      required:
        - job_id
        - instrument_id
        - instrument_type
        - valuation_date
        - status
        - total_fair_value
        - components
        - key_parameters
        - reproducibility
        - warnings
        - errors
      properties:
        job_id: { type: integer }
        instrument_id: { type: integer }
        instrument_type: { $ref: '#/components/schemas/InstrumentType' }
        valuation_date: { type: string, format: date }
        status: { type: string, enum: [DONE, FAILED, PARTIAL] }
        total_fair_value: { type: [number, 'null'] }
        per_unit_value: { type: [number, 'null'] }
        components: { $ref: '#/components/schemas/Components' }
        key_parameters: { $ref: '#/components/schemas/KeyParameters' }
        sensitivity_summary: { type: [object, 'null'] }
        scenario_summary: { type: [object, 'null'] }
        audit_data: { $ref: '#/components/schemas/AuditData' }
        reproducibility: { $ref: '#/components/schemas/Reproducibility' }
        warnings:
          type: array
          default: []
          items: { $ref: '#/components/schemas/Issue' }
        errors:
          type: array
          default: []
          items: { $ref: '#/components/schemas/Issue' }

    # 표준 12종 key 만 허용. total_fair_value 는 나머지 부호포함 합의 echo.
    Components:
      type: object
      additionalProperties: false
      properties:
        bond_value: { type: [number, 'null'], description: '채권(host), 부호 +' }
        preferred_share_value: { type: [number, 'null'], description: '우선주(host), 부호 +' }
        conversion_option_value: { type: [number, 'null'], description: '전환옵션, 부호 +' }
        exchange_option_value: { type: [number, 'null'], description: '교환옵션, 부호 +' }
        warrant_value: { type: [number, 'null'], description: '신주인수권, 부호 +' }
        redemption_option_value: { type: [number, 'null'], description: '상환권(풋), 부호 +' }
        issuer_call_value: { type: [number, 'null'], description: '발행자 콜, 부호 −' }
        sale_claim_value: { type: [number, 'null'], description: '매도청구권, 부호 −' }
        stock_option_value: { type: [number, 'null'], description: '주식매수선택권, 부호 +' }
        conditional_option_value: { type: [number, 'null'], description: '조건부 SO, 부호 +' }
        dilution_effect: { type: [number, 'null'], description: '희석효과, 부호 −' }
        total_fair_value: { type: [number, 'null'], description: '= 나머지 부호포함 합의 echo' }
    KeyParameters:
      type: object
      additionalProperties: false
      required: [model_name, model_version]
      properties:
        risk_free_rate: { type: [number, 'null'] }
        ytm: { type: [number, 'null'] }
        credit_spread: { type: [number, 'null'] }
        volatility: { type: [number, 'null'] }
        dividend_yield: { type: [number, 'null'] }
        parity: { type: [number, 'null'], description: '엔진 산출. 입력 아님(패치 2.1).' }
        discount_rate: { type: [number, 'null'] }
        model_name: { type: string }
        model_version: { type: string }
        simulation_paths: { type: [integer, 'null'] }
        lattice_steps: { type: [integer, 'null'] }
    Reproducibility:
      type: object
      required: [input_hash, seed, model_version]
      properties:
        input_hash: { type: string, pattern: '^[a-f0-9]{64}$' }
        seed: { type: integer }
        model_version: { type: string }
        engine_commit: { type: [string, 'null'] }
        computed_at: { type: string, format: date-time }

    # audit-data 응답 + PricingResult.audit_data 공용.
    AuditData:
      type: [object, 'null']
      properties:
        cashflow_schedule: { type: array, items: { type: object, additionalProperties: true } }
        discount_factors: { type: array, items: { type: object, additionalProperties: true } }
        curve_snapshot: { type: object, additionalProperties: true }
        node_summary: { type: object, additionalProperties: true }
        path_summary: { type: object, additionalProperties: true }
        exercise_log: { type: array, items: { type: object, additionalProperties: true } }
        calculation_trace: { type: array, items: { type: object, additionalProperties: true } }
        input_snapshot_hash: { type: string }